refactored model into services
